add brand delivery method, distance and min consumption
